refactor POM hierarchy
