+++ app/components/ui/combobox.tsx
@@ -100,7 +100,7 @@ export default function Combobox({ isSelectGraph = false, disabled = false, inTa
           </TooltipContent>
         </Tooltip>
         <SelectContent className="min-w-52 max-h-[40lvh] bg-foreground">
-          <div className="p-4">
+          <div className="p-4" id="graphSearch">
           <Input ref={ref => ref?.focus()} className="w-full" placeholder={`Search a graph ${type}`} onChange={(e) => {
             setSearch(e.target.value)
             setMaxOptions(5)
