+++ e2e/config/constants.ts
@@ -1,0 +1,7 @@
+import queryData from '../config/queries.json'
+
+export const BATCH_CREATE_PERSONS = queryData.queries[0].testQueries[0].query;
+export const FETCH_FIRST_TEN_NODES = queryData.queries[0].testQueries[1].query;
+export const BATCH_CREATE_PERSONS_APIREQ = queryData.queries[0].testQueries[0].apiReq;
+export const CREATE_SAMPLE_PERSONS_QUERY_APIREQ = queryData.queries[0].testQueries[2].apiReq;
+export const FETCH_NODE_DETAILS_QUERY = queryData.queries[0].testQueries[3].query;
